backend/pyproject.toml
[project]
name = "template-vue3-django-ninja-backend"
version = "0.1.0"
description = "Vue3 + Django-Ninja 后端 API 服务"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<5.1",
    "django-ninja>=1.1,<1.2",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.0",
    "pyjwt>=2.8.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "bcrypt>=4.1.0",
    "django-redis>=5.4.0",
    "django-ratelimit>=4.1.0",
]

[project.optional-dependencies]
production = [
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=4.3.0",
    "ipython>=8.21.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
    "factory-boy>=3.3.0",
    "freezegun>=1.4.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
python_files = ["test_*.py", "*_test.py"]
